+++ imj-game-hamazed/src/Imj/Game/Hamazed/Network/Internal/Types.hs
@@ -51,7 +51,7 @@ import           Imj.Game.Level
 import           Imj.Game.Status
 import           Imj.Game.Hamazed.Timing
 import           Imj.Game.Hamazed.Music
-import           Imj.Music
+import           Imj.Music.Types
 import           Imj.Timing
 
 -- | An event generated by the client, sent to the server.
