+++ imj-game/src/Imj/Game/Audio/Class.hs
@@ -7,7 +7,7 @@ module Imj.Game.Audio.Class
 import           Imj.Prelude
 import           Control.Monad.IO.Class(MonadIO)
 import           Control.Monad.IO.Unlift(MonadUnliftIO)
-import           Imj.Music
+import           Imj.Music.Types
 
 class Audio e where
   -- | The value to use when the user didn'T specify one on the command line.
